chore(cloudfront-sign): rename source file
