+++ packages/cloudfront-request-presigner/src/sign.ts
